Move KVConnectorOutput to outputs.py
