+++ tests/v1/kv_connector/unit/test_remote_prefill_lifecycle.py
@@ -2,9 +2,8 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 import copy
 
-from vllm.v1.outputs import EMPTY_MODEL_RUNNER_OUTPUT
+from vllm.v1.outputs import EMPTY_MODEL_RUNNER_OUTPUT, KVConnectorOutput
 from vllm.v1.request import FinishReason, RequestStatus
-from vllm.v1.worker.kv_connector_model_runner_mixin import KVConnectorOutput
 
 from .utils import (assert_scheduler_empty, create_model_runner_output,
                     create_request, create_scheduler, create_vllm_config)
